+++ packages/alfa-array/src/array.ts
@@ -6,6 +6,7 @@ import { Serializable } from "@siteimprove/alfa-json";
 import { Iterable } from "@siteimprove/alfa-iterable";
 import { None, Option, Some } from "@siteimprove/alfa-option";
 import { Predicate } from "@siteimprove/alfa-predicate";
+import { Reducer } from "@siteimprove/alfa-reducer";
 import { Refinement } from "@siteimprove/alfa-refinement";
 
 import * as global from "./global";
@@ -56,6 +57,18 @@ export namespace Array {
 
   export function clone<T extends Clone<T>>(array: Array<T>): Array<T> {
     return array.map(Clone.clone);
+  }
+
+  export function reduce<T, U = T>(
+    array: Array<T>,
+    reducer: Reducer<T, U, [number]>,
+    accumulator: U
+  ): U {
+    for (let i = 0, n = array.length; i < n; i++) {
+      accumulator = reducer(accumulator, array[i], i);
+    }
+
+    return accumulator;
   }
 
   export function find<T, U extends T>(